Added trained models and models code
The savedModels folders have saved models that you can load in the python code on lines 96 and 99 as of commit. They are encoders trained on the abundance and/or marker profile. Also, there is a perceptron at the end but it sucks and was not really worth saving at the moment. It could be the encoders or the perceptron that are not effective. In the paper, the convolutional autoencoder with random forest output performed the best for the colorectal dataset. The marker autoencoders "me" are too large for github but can be trained and stored locally.